% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_word.R
\name{add_word}
\alias{add_word}
\title{Copies characters of the word into the puzzle_array}
\usage{
add_word(puzzle_array, word, location, offset)
}
\arguments{
\item{puzzle_array}{- character vector containing the puzzle}

\item{word}{1 element character vector containing the word being placed
in the puzzle}

\item{location}{row and column location}

\item{offset}{number of rows and columns the location is to be offset.}
}
\description{
Copies characters of the word into the puzzle_array
}
