% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{case}
\alias{case}
\title{Returns the word either without change of folded to upper case depending on
the second argument of class logical.}
\usage{
case(word, upper_case)
}
\arguments{
\item{word}{1 element character vector with the word}

\item{upper_case}{a logical flag indicating whether or not the word is
folded to upper case or not.}
}
\description{
Returns the word either without change of folded to upper case depending on
the second argument of class logical.
}
