% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_word.R
\name{collapse_word}
\alias{collapse_word}
\title{Returns the character vector collapsed into as a single character vector 
element.}
\usage{
collapse_word(x, collapse = "")
}
\arguments{
\item{x}{character vector to be collapsed}

\item{collapse}{charcter string with collapse argument.}
}
\description{
Returns the character vector collapsed into as a single character vector 
element.
}
