% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_word.R
\name{expand_word}
\alias{expand_word}
\title{Returns a list with each character from the word in seccessive list elements}
\usage{
expand_word(word)
}
\arguments{
\item{word}{a 1 element character vector contain a word to be expanded}
}
\description{
Returns a list with each character from the word in seccessive list elements
}
