% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_out_bad_directions.R
\name{filter_out_bad_directions}
\alias{filter_out_bad_directions}
\title{Returns a vector of directions that can be used
given the word size, the demensions of the puzzle, and the proposed list of
directions.}
\usage{
filter_out_bad_directions(word, directions, specifications)
}
\arguments{
\item{word}{- character vector of one element with word to be put into 
puzzle array.}

\item{directions}{- character vector of the directions proposed for use.}

\item{specifications}{- puzzle specifications}
}
\description{
Returns a vector of directions that can be used
given the word size, the demensions of the puzzle, and the proposed list of
directions.
}
