% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_words.R
\name{find_words}
\alias{find_words}
\title{Prints out location for each \emph{word} in a character vector}
\usage{
find_words(words, text, found_words)
}
\arguments{
\item{words}{charcter vector of words}

\item{text}{character vector of length 1 that is being searched}

\item{found_words}{dataframe of words found with three columns:
\emph{n_word} -- integer order of the word found, 
\emph{original_word} -- the word searched for and found with original case
retained.
\emph{starts_at} -- characer position of the first letter of the found
word within the \emph{text} provided.}
}
\description{
Prints out location for each \emph{word} in a character vector
}
