% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_puzzle_array.R
\name{form_puzzle_array}
\alias{form_puzzle_array}
\title{Forms the empty puzzle array with  `.' as the holder in each cell.
The dimensions are determined by the specifications.}
\usage{
form_puzzle_array(specifications)
}
\arguments{
\item{specifications}{- puzzle specifications}
}
\description{
Forms the empty puzzle array with  `.' as the holder in each cell.
The dimensions are determined by the specifications.
}
