% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/found_forward_and_backward_word_location.R
\name{found_forward_and_backward_word_location}
\alias{found_forward_and_backward_word_location}
\title{Looks for words in text and returns found words listed forward and backward}
\usage{
found_forward_and_backward_word_location(words, text)
}
\arguments{
\item{words}{charcter vector of words}

\item{text}{character vector of length 1 that is being searched}
}
\value{
dataframe with words listed in orginal and reverse orders
}
\description{
Finds as many words as possible with in text and lists them in forward
and backword.
}
