% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_directions.R
\name{get_random_directions}
\alias{get_random_directions}
\title{Returns a character vector of directs with \emph{n_directions} elements,
where \emph{n_directions} is from the \emph{specifications}. All directions
are allowed by the word size and the puzzle_array dimensions.}
\usage{
get_random_directions(word, specifications)
}
\arguments{
\item{word}{- character vector of one element with word to be put into 
puzzle array.}

\item{specifications}{- puzzle specifications}
}
\description{
Returns a character vector of directs with \emph{n_directions} elements,
where \emph{n_directions} is from the \emph{specifications}. All directions
are allowed by the word size and the puzzle_array dimensions.
}
