% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specifications.R
\name{get_specifications}
\alias{get_specifications}
\title{Reads the puzzle description file and returns a character
vector with one specification in each element.}
\usage{
get_specifications(puzzle_description_file)
}
\arguments{
\item{puzzle_description_file}{a character varialble containing the 
name of the file containing the word list and the specifications, which
have the following elements:}
}
\description{
Reads the puzzle description file and returns a character
vector with one specification in each element.
}
