% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text_string.R
\name{get_text_string}
\alias{get_text_string}
\title{Concatenates and removes all blanks and punctuation all words in each line.}
\usage{
get_text_string(text_file)
}
\arguments{
\item{text_file}{file containing text to have spaces and punctuation 
removed and words concatenated removing any white space.}
}
\description{
Concatenates and removes all blanks and punctuation all words in each line.
}
