% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_location.R
\name{get_trial_location}
\alias{get_trial_location}
\title{Returns a two element numeric vector with the row and column specified
for a location to be tried.}
\usage{
get_trial_location(word, direction, specifications)
}
\arguments{
\item{word}{- character vector of one element with word to be put into 
puzzle array.}

\item{direction}{- character vector of one element that specifies the 
proposed direction.}

\item{specifications}{- puzzle specifications}
}
\description{
Returns a two element numeric vector with the row and column specified
for a location to be tried.
}
