% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_word_list.R
\name{get_trial_word_list}
\alias{get_trial_word_list}
\title{Returns a list containing the following:
   trial_word_vector - character vector containing the contents of the 
       puzzle_array beginning at the location selected and continuing
       for the length of the word and in the direction indicated.
   location - 2 element integer vector containing a random location of the 
       first character of the word
   direction - 1 element character vector containing the direction
       descriptor}
\usage{
get_trial_word_list(puzzle_array, word, direction, specifications)
}
\arguments{
\item{puzzle_array}{character vector containing the puzzle}

\item{word}{character vector of one element with word to be put into 
puzzle array.}

\item{direction}{character vector of one element that specifies the 
proposed direction.}

\item{specifications}{puzzle specifications}
}
\description{
Returns a list containing the following:
   trial_word_vector - character vector containing the contents of the 
       puzzle_array beginning at the location selected and continuing
       for the length of the word and in the direction indicated.
   location - 2 element integer vector containing a random location of the 
       first character of the word
   direction - 1 element character vector containing the direction
       descriptor
}
