% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_word_order.R
\name{get_word_order}
\alias{get_word_order}
\title{Orders words in decreasing length. Ties are in ASCII order.}
\usage{
get_word_order(words)
}
\arguments{
\item{words}{Character vector with a word in each element.}
}
\description{
Orders words in decreasing length. Ties are in ASCII order.
}
