% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_words.R
\name{get_words}
\alias{get_words}
\title{get_words reads in words to be inserted in the puzzle and the specifications of the puzzle.}
\usage{
get_words(puzzle_description_file)
}
\arguments{
\item{puzzle_description_file}{is a character varialble containing the 
name of the file containing the word list and specifications.

It has a list of words ended by a line beginning with a '#'. 
Words may be on the same or different lines.
The following elements make of the specifications, which follow the word
list:
\tabular{ll}{
puzzle_description_file \tab File name (with path if necessary) with the specifications and word list to use. \cr
title \tab Title for word puzzle first page.\cr
max_trials \tab Number of times the program is to try to fit each word into the puzzle \emph{n_direction} times.\cr
n_directions \tab Number of directions (not necessarily different) to be tried for each word if needed. \cr
width \tab Width of puzzle in characters. \cr
height \tab Height of puzzle in rows. \cr
left_right \tab Direction of word characters; normal. \cr
right_left \tab Direction of word characters; backward. \cr
up_down \tab Direction of word characters \cr
down_up \tab Direction of word characters \cr
up_right \tab Direction of word characters \cr
up_left \tab Direction of word characters \cr
down_right \tab Direction of word characters \cr
down_left \tab Direction of word characters \cr
}}
}
\description{
Reads the puzzle description file and returns a character
vector with one word in each element.
}
\details{
The letter case of words and order is not changed.
}
