% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_puzzle.R
\name{make_puzzle}
\alias{make_puzzle}
\title{Returns a list with the filename, the specifications, a puzzle_array, and 
the word_key.}
\usage{
make_puzzle(filename, upper_case = TRUE)
}
\arguments{
\item{filename}{the name of the file with the puzzle words and specifications}

\item{upper_case}{a logical flag indicating whether or not the puzzle is
folded to upper case or not.}
}
\description{
Returns a list with the filename, the specifications, a puzzle_array, and 
the word_key.
}
