% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_key.R
\name{print_key}
\alias{print_key}
\title{Creates puzzle text ready to print}
\usage{
print_key(puzzle_list, key = "both")
}
\arguments{
\item{puzzle_list}{list containing list of words to use and the puzzle 
specifications}

\item{key}{can have values of \emph{both} or \emph{row}. If \emph{both} 
then row and column location will be provided, else just the row is 
provided.}
}
\description{
Creates puzzle text ready to print
}
