% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_puzzle.R
\name{print_puzzle}
\alias{print_puzzle}
\title{Creates puzzle text ready to print}
\usage{
print_puzzle(puzzle_list)
}
\arguments{
\item{puzzle_list}{list containing list of words to use and the puzzle 
specifications}
}
\description{
Creates puzzle text ready to print
}
