% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_words.R
\name{print_words}
\alias{print_words}
\title{Prints list of words to find}
\usage{
print_words(puzzle_list)
}
\arguments{
\item{puzzle_list}{list containing list of words to use and the puzzle}
}
\description{
Prints list of words to find
}
