% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_word_fit.R
\name{try_word_fit}
\alias{try_word_fit}
\title{Returns a list consisting of a logical value indicating whether or not
the proposed word fits the puzzle array, a two integer vector of the row
and column location of the first character, and a two integer vector that
determines the direction of the word's orientation.
try_word_fit() - the characters found within
the array, an row and column location in a two integer vector}
\usage{
try_word_fit(puzzle_array, word, direction, specifications)
}
\arguments{
\item{puzzle_array}{character vector containing the puzzle}

\item{word}{1 element character vector containing the word being placed
in the puzzle}

\item{direction}{character vector of one element that specifies the 
proposed direction.}

\item{specifications}{puzzle specifications}
}
\description{
Returns a list consisting of a logical value indicating whether or not
the proposed word fits the puzzle array, a two integer vector of the row
and column location of the first character, and a two integer vector that
determines the direction of the word's orientation.
try_word_fit() - the characters found within
the array, an row and column location in a two integer vector
}
